classrooms/requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.6.0.0
blessings==1.6.1
bpython==0.17
celery==4.3.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
curtsies==0.2.11
dataclasses==0.6
django~=2.1.7
django-appconf==1.0.3
django-compressor==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
django-mysql==2.4.1
django-recaptcha2==1.4.1
django-sass-processor==0.7.2
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djongo==1.2.32
Faker==0.8.9
flower==0.9.3
greenlet==0.4.12
idna==2.8
kombu==4.4.0
libsass==0.17.0
mysql==0.0.2
mysqlclient==1.4.1
Pillow==5.4.1
protobuf==3.7.0rc2
psycopg2==2.7.7
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
pymongo==3.7.2
python-axolotl==0.1.42
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.0
pytz==2018.9
rcssmin==1.0.6
redis==3.2.1
requests==2.21.0
requests-cache==0.4.13
rjsmin==1.0.12
selenium==3.141.0
six==1.12.0
South==1.0.2
sportmonks==1.0.0
sqlparse==0.2.4
text-unidecode==1.1
tornado==5.1.1
tzlocal==1.5.1
ua-parser==0.3.6
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webwhatsapi==2.0.5
django-jet
data_wizard
PyPDF2
reportlab
svglib
psycopg2-binary
zeep
bs4